refactor(tooling): centralize positive env integers
