+++ scripts/lib/numeric-options.d.mts
@@ -1,3 +1,4 @@
 export function parsePositiveInt(raw: string, label: string): number;
+export function readPositiveEnvInt(name: string, env: NodeJS.ProcessEnv, fallback: number): number;
 export function parseNonNegativeInt(raw: string, label: string): number;
 export function parsePositiveNumber(raw: string, label: string): number;
