+++ scripts/lib/numeric-options.mjs
@@ -12,6 +12,22 @@ export function parsePositiveInt(raw, label) {
   return value;
 }
 
+/** Read a safe positive integer from an environment variable. */
+export function readPositiveEnvInt(name, env, fallback) {
+  const raw = env[name]?.trim();
+  if (raw === undefined || raw === "") {
+    return fallback;
+  }
+  if (!/^[1-9]\d*$/u.test(raw)) {
+    throw new Error(`invalid ${name}: ${raw}`);
+  }
+  const value = Number(raw);
+  if (!Number.isSafeInteger(value)) {
+    throw new Error(`invalid ${name}: ${raw}`);
+  }
+  return value;
+}
+
 /** Parse a safe non-negative integer option. */
 export function parseNonNegativeInt(raw, label) {
   const text = String(raw).trim();
